Applied x64 patch
Using patch from http://scobyseo.blogspot.com/2009/12/disksim-40-ssd-extention-64bit-patch.html
but only for disksim without SSD support.

+++ src/disksim_iosim.c
@@ -354,7 +354,7 @@ fprintf (outputfile, "devno %d, depth %d, currbus %d, inslotno %d, master %d\n",
 
 
 
-static int iosim_load_map(struct lp_block *b, int n) {
+static int iosim_load_map(struct lp_block *b, int64_t n) {
   int c;
   int i = 0;
   char *s = 0;
